Added the ability to change priority of outline
Items in the hitscore mode's list can be either Miss/Hitscore/Angle/Accuracy. I've added the ability to sort through the priority of these (so you can have acc only take priority over angle colouring for example). This was possible in config by changing the order of the list but it can now be done ingame.